Move hs2 code to arrow/dbi
