+++ cpp/src/arrow/dbi/hiveserver2/api.h
@@ -17,20 +17,11 @@
 
 #pragma once
 
-#include <string>
+#include "arrow/dbi/hiveserver2/columnar-row-set.h"
+#include "arrow/dbi/hiveserver2/operation.h"
+#include "arrow/dbi/hiveserver2/service.h"
+#include "arrow/dbi/hiveserver2/session.h"
+#include "arrow/dbi/hiveserver2/types.h"
+#include "arrow/dbi/hiveserver2/util.h"
 
-#include "arrow/adapters/hiveserver2/operation.h"
-
-namespace arrow {
-namespace hiveserver2 {
-
-// Utility functions. Intended primary for testing purposes - clients should not
-// rely on stability of the behavior or API of these functions.
-class Util {
- public:
-  // Fetches the operation's results and returns them in a nicely formatted string.
-  static void PrintResults(const Operation* op, std::ostream& out);
-};
-
-}  // namespace hiveserver2
-}  // namespace arrow
+#include "arrow/status.h"
