+++ cpp/src/arrow/dbi/hiveserver2/columnar-row-set.h
