+++ cpp/src/arrow/dbi/hiveserver2/hiveserver2-test.cc
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/adapters/hiveserver2/operation.h"
+#include "arrow/dbi/hiveserver2/operation.h"
 
 #include <memory>
 #include <sstream>
@@ -24,9 +24,9 @@
 
 #include <gtest/gtest.h>
 
-#include "arrow/adapters/hiveserver2/service.h"
-#include "arrow/adapters/hiveserver2/session.h"
-#include "arrow/adapters/hiveserver2/thrift-internal.h"
+#include "arrow/dbi/hiveserver2/service.h"
+#include "arrow/dbi/hiveserver2/session.h"
+#include "arrow/dbi/hiveserver2/thrift-internal.h"
 
 #include "arrow/status.h"
 #include "arrow/test-util.h"
