+++ cpp/src/arrow/dbi/hiveserver2/operation.cc
@@ -15,12 +15,12 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/adapters/hiveserver2/operation.h"
+#include "arrow/dbi/hiveserver2/operation.h"
 
-#include "arrow/adapters/hiveserver2/thrift-internal.h"
+#include "arrow/dbi/hiveserver2/thrift-internal.h"
 
-#include "arrow/adapters/hiveserver2/ImpalaService_types.h"
-#include "arrow/adapters/hiveserver2/TCLIService.h"
+#include "arrow/dbi/hiveserver2/ImpalaService_types.h"
+#include "arrow/dbi/hiveserver2/TCLIService.h"
 
 #include "arrow/status.h"
 #include "arrow/util/logging.h"
