+++ cpp/src/arrow/dbi/hiveserver2/public-api-test.cc
