+++ cpp/src/arrow/dbi/hiveserver2/service.h
