+++ cpp/src/arrow/dbi/hiveserver2/session.cc
@@ -15,10 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/adapters/hiveserver2/session.h"
+#include "arrow/dbi/hiveserver2/session.h"
 
-#include "arrow/adapters/hiveserver2/TCLIService.h"
-#include "arrow/adapters/hiveserver2/thrift-internal.h"
+#include "arrow/dbi/hiveserver2/TCLIService.h"
+#include "arrow/dbi/hiveserver2/thrift-internal.h"
 
 #include "arrow/status.h"
 #include "arrow/util/logging.h"
