+++ cpp/src/arrow/dbi/hiveserver2/thrift/.gitignore
