+++ cpp/src/arrow/dbi/hiveserver2/thrift/ExecStats.thrift
