+++ cpp/src/arrow/dbi/hiveserver2/thrift/ImpalaService.thrift
