+++ cpp/src/arrow/dbi/hiveserver2/thrift/Status.thrift
