+++ cpp/src/arrow/dbi/hiveserver2/thrift/TCLIService.thrift
