+++ cpp/src/arrow/dbi/hiveserver2/thrift/Types.thrift
