+++ cpp/src/arrow/dbi/hiveserver2/thrift/beeswax.thrift
