+++ cpp/src/arrow/dbi/hiveserver2/thrift/fb303.thrift
