+++ cpp/src/arrow/dbi/hiveserver2/thrift/generate_error_codes.py
