+++ cpp/src/arrow/dbi/hiveserver2/thrift/hive_metastore.thrift
