+++ cpp/src/arrow/dbi/hiveserver2/types.cc
@@ -15,9 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/adapters/hiveserver2/types.h"
+#include "arrow/dbi/hiveserver2/types.h"
 
-#include "arrow/adapters/hiveserver2/thrift-internal.h"
+#include "arrow/dbi/hiveserver2/thrift-internal.h"
 
 #include "arrow/util/logging.h"
