+++ cpp/src/arrow/dbi/hiveserver2/types.h
