+++ cpp/src/arrow/dbi/hiveserver2/util.cc
@@ -15,18 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/adapters/hiveserver2/util.h"
+#include "arrow/dbi/hiveserver2/util.h"
 
 #include <algorithm>
 #include <memory>
 #include <sstream>
 #include <vector>
 
-#include "arrow/adapters/hiveserver2/columnar-row-set.h"
-#include "arrow/adapters/hiveserver2/thrift-internal.h"
+#include "arrow/dbi/hiveserver2/columnar-row-set.h"
+#include "arrow/dbi/hiveserver2/thrift-internal.h"
 
-#include "arrow/adapters/hiveserver2/TCLIService.h"
-#include "arrow/adapters/hiveserver2/TCLIService_types.h"
+#include "arrow/dbi/hiveserver2/TCLIService.h"
+#include "arrow/dbi/hiveserver2/TCLIService_types.h"
 
 #include "arrow/status.h"
